requirements.txt
# ------------------------------
# FASTAPI + SERVER
# ------------------------------
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1

# ------------------------------
# DATABASE (MongoDB)
# ------------------------------
motor==3.6.0
pymongo[srv]==4.9.2
dnspython==2.6.1

# ------------------------------
# SECURITY / AUTH
# ------------------------------
passlib==1.7.4
python-jose[cryptography]==3.3.0

# ------------------------------
# AZURE SERVICES
# ------------------------------
azure-storage-blob==12.22.0
azure-core==1.30.0
azure-identity==1.17.1

# OpenAI SDK novo (suporta AzureOpenAI 2024+)
openai==1.55.3  

# ------------------------------
# UTILITÁRIOS
# ------------------------------
requests==2.32.3
pydantic==2.9.2
annotated-types==0.7.0
typing_extensions==4.12.2
moviepy==1.0.3
email-validator==2.1.0.post1
pillow==9.5.0